% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_post.R
\name{as_post_array}
\alias{as_post_array}
\alias{as_post_array.sf}
\title{Create a post_array object}
\usage{
as_post_array(
  x,
  group_id = NULL,
  time_column_name = NULL,
  sf_column_name = NULL,
  geometry_summary = "centroid"
)

\method{as_post_array}{sf}(
  x,
  group_id = NULL,
  time_column_name = NULL,
  sf_column_name = NULL,
  geometry_summary = "centroid"
)
}
\arguments{
\item{x}{object to convert to \code{post_array} with \code{POLYGON}/\code{MULTIPOLYGON}
geometries and a date/time column.}

\item{group_id}{(character) name of column containing the grouping
identifier for the changing polygons or group identifier given as an
integer or character value that is assigned to all rows in \code{x}.
Defaults to the first column in \code{x}.}

\item{time_column_name}{(character) name of column with the temporal
dimension information. Defaults to the first temporal column in \code{x}.}

\item{sf_column_name}{(character) name of column with changing
geometries. Defaults to active \code{sf_column}.}

\item{geometry_summary}{(character) function to compute the summary
geometry. Supported arguments are \code{union} to compute the union
and dissolve of the changing geometries per group, \code{centroid} to
compute the centroid of the union and dissolve per group, and
\code{bbox} to compute the bbox of the union and dissolve per group.}
}
\value{
an object of class \code{post_array}.
}
\description{
\code{post_array} is a spatio-temporal array data structure that
organises polygons that change their shape in space and time.
It extends the \code{stars} class for vector data cubes to support
changing geometries as attributes.
The \code{post_array} class supports two dimensions: \code{geom_sum} and
\code{datetime}. \code{geom_sum} is the summary geometry of the changing
shapes of the polygon geometries.
}
\examples{
as_post_array(polygons, "gid", "datetime")

if(require(dplyr, quietly = TRUE)) {
  library(sf, quietly = TRUE)
  polygons |>
    mutate(area = st_area(geometry)) |>
    as_post_array("gid", "datetime", geometry_summary = "union")
}

}
\references{
Abad, L., Sudmanns, M., and Hölbling, D. (2024)
Vector data cubes for features evolving in space and time,
\emph{AGILE GIScience Ser.}, 5, 16,
\url{https://doi.org/10.5194/agile-giss-5-16-2024}
}
