% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot}
\alias{autoplot}
\alias{autoplot.post_array}
\alias{autoplot.post_table}
\title{Plot post_array or post_table with ggplot2}
\usage{
autoplot.post_array(
  object,
  attribute = NULL,
  max_plot = 9,
  geom_sum = FALSE,
  ...
)

autoplot.post_table(
  object,
  attribute = NULL,
  max_plot = 9,
  geom_sum = FALSE,
  ...
)
}
\arguments{
\item{object}{a post_array or post_table object}

\item{attribute}{the attribute to fill the geometries}

\item{max_plot}{maximum number of subplots rendered}

\item{geom_sum}{should the geom_sum geometries be used for plotting? Defaults to \code{FALSE}}

\item{...}{ignored}
}
\value{
an object of class \code{ggplot}
}
\description{
Plot the changing or summary geometries in a post_array or post_table as facets
defined by the first non spatial dimension. Other dimensions are discarded.
}
\details{
See \code{ggplot2::autoplot()}.
}
\examples{
library(ggplot2)
library(sf)
arr = as_post_array(polygons, geometry_summary = summarise_geometry_union)
arr$area = st_area(arr$geometry)
autoplot(arr, area)
autoplot(arr, area, geom_sum = TRUE)
autoplot(arr, area, max_plot = 3)

tab = as_post_table(polygons, geometry_summary = summarise_geometry_union)
autoplot(tab, gid)
autoplot(tab, gid, geom_sum = TRUE)

}
