% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{binary-predicates-post-array}
\alias{binary-predicates-post-array}
\alias{st_mask}
\title{sf geometry predicates for post_array objects}
\usage{
st_mask(x, y, .predicate = st_intersects, ...)
}
\arguments{
\item{x}{a post_array object}

\item{y}{an object of class \code{sf}, \code{sfc}, \code{sfg}}

\item{.predicate}{binary predicate to use (see \code{?sf::geos_binary_pred}), defaults to \code{st_intersects}}

\item{...}{passed on to the geometry predicate}
}
\value{
a one column matrix as the result of any features in y relating to
any feature in x
}
\description{
sf geometry predicates for post_array objects
}
\examples{
library(sf)
arr = as_post_array(polygons, geometry_summary = summarise_geometry_union)
circles = list(st_point(c(0.55,0.5)), st_point(c(0.75,0.4))) |>
  st_sfc(crs = 4326) |>
  st_buffer(10000)
st_mask(arr, circles)
arr$circles_disjoint = st_mask(arr, circles, .predicate = st_disjoint)
}
