% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_group_id}
\alias{check_group_id}
\title{Check and assign defaults to post object arguments}
\usage{
check_group_id(x = x, group_id = NULL)
}
\description{
Check and assign defaults to post object arguments
}
