% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{classes}
\alias{classes}
\alias{remove_post_table}
\alias{restore_post_table}
\alias{restore_spatial_post_table}
\alias{restore_temporal_post_table}
\alias{remove_post_array}
\alias{restore_post_array}
\title{Remove or restore post_* classes}
\usage{
remove_post_table(x)

restore_post_table(x)

restore_spatial_post_table(x)

restore_temporal_post_table(x)

remove_post_array(x)

restore_post_array(x, x_orig)
}
\arguments{
\item{x}{An object that inherits stars class.}

\item{x_orig}{An object that inherits post_array class}
}
\value{
An object that inherits or removes the classes post_table/post_array
and sf.
}
\description{
Utility functions to remove/restore post_table/post_array classes when
passing functions onto stars, spatial_cubble_df and temporal_cubble_df
}
