% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble.R
\name{cubble-attrs}
\alias{cubble-attrs}
\alias{spatial.post_table}
\title{Extract cubble attributes}
\usage{
\method{spatial}{post_table}(data)
}
\arguments{
\item{data}{a post_table object}
}
\value{
a post_table object
}
\description{
Extract cubble attributes
}
