% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble.R
\name{cubble-dplyr}
\alias{cubble-dplyr}
\alias{arrange.post_table}
\alias{filter.post_table}
\title{dplyr methods for post_table}
\usage{
\method{arrange}{post_table}(.data, ...)

\method{filter}{post_table}(.data, ...)
}
\arguments{
\item{...}{see corresponding dplyr function}

\item{data, .data}{a post_table object}
}
\description{
See \code{cubble::dplyr()} for details
}
