% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble.R
\name{cubble-face}
\alias{cubble-face}
\alias{face_spatial.post_table}
\alias{face_temporal.post_table}
\title{Face spatial/temporal method for post_table}
\usage{
\method{face_spatial}{post_table}(data)

\method{face_temporal}{post_table}(data, col)
}
\arguments{
\item{data}{a post_table object}

\item{col}{see \code{cubble::face_temporal()} for details.}
}
\value{
a post_table object
}
\description{
\code{face_spatial} pivots a \code{post_table} object from a long (temporal)
to a nested (spatial) form.
\code{face_temporal()} pivots a \code{post_table} object from a nested (spatial)
to a long (temporal) form.
For \code{post_table} objects, it activates the changing geometry
in the long form.
}
