% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble.R
\name{cubble-fill}
\alias{cubble-fill}
\alias{fill_gaps.post_table}
\title{Gap-filling on cubble}
\usage{
\method{fill_gaps}{post_table}(.data, ..., .full, .start, .end)
}
\arguments{
\item{.data}{a post_table object}

\item{...}{see \code{?tsibble::fill_gaps}}

\item{.full}{see \code{?tsibble::fill_gaps}}

\item{.start}{see \code{?tsibble::fill_gaps}}

\item{.end}{see \code{?tsibble::fill_gaps}}
}
\value{
a post_table object
}
\description{
Gap-filling on cubble
}
