% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKM.R
\name{distanceMatrix}
\alias{distanceMatrix}
\title{distanceMatrix}
\usage{
distanceMatrix(tree, OTU)
}
\arguments{
\item{tree}{the phylogenetic tree}

\item{OTU}{the OTU table}
}
\description{
distanceMatrix
}
\keyword{internal}
