% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr-post-array}
\alias{dplyr-post-array}
\alias{filter.post_array}
\alias{mutate.post_array}
\alias{rename.post_array}
\alias{select.post_array}
\alias{slice.post_array}
\alias{transmute.post_array}
\title{dplyr methods for post_array}
\usage{
\method{filter}{post_array}(.data, ...)

\method{mutate}{post_array}(.data, ...)

\method{rename}{post_array}(.data, ...)

\method{select}{post_array}(.data, ...)

\method{slice}{post_array}(.data, ..., drop = FALSE)

\method{transmute}{post_array}(.data, ...)
}
\arguments{
\item{...}{see corresponding function in package dplyr}

\item{data, .data}{a post_array object}
}
\description{
dplyr methods for post_array
}
\details{
See \code{?stars::dplyr} for details
}
