% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr-post-table}
\alias{dplyr-post-table}
\alias{dplyr_reconstruct.post_table}
\alias{arrange.post_table}
\alias{filter.post_table}
\alias{mutate.post_table}
\alias{rename.post_table}
\alias{select.post_table}
\alias{slice.post_table}
\title{dplyr methods for post_table}
\usage{
\method{dplyr_reconstruct}{post_table}(data, template)

\method{arrange}{post_table}(.data, ...)

\method{filter}{post_table}(.data, ...)

\method{mutate}{post_table}(.data, ...)

\method{rename}{post_table}(.data, ...)

\method{select}{post_table}(.data, ...)

\method{slice}{post_table}(.data, ..., .dots)
}
\arguments{
\item{data, .data, x}{a post_table object}

\item{template}{Template data frame to use for restoring attributes.}

\item{...}{see corresponding function in package dplyr}

\item{.dots}{see corresponding function in package \code{dplyr}}
}
\description{
dplyr methods for post_table
}
\details{
See \code{??cubble::dplyr} for details

See \code{?sf::tidyverse} for details
}
