% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKM.R
\name{getkernel}
\alias{getkernel}
\title{getkernel}
\usage{
getkernel(dv, h, OTU)
}
\arguments{
\item{dv}{distance matrix vector}

\item{OTU}{the OTU table}

\item{hx}{h value, c * sd}
}
\description{
getkernel
}
\keyword{internal}
