% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKM.R
\name{getkernellist}
\alias{getkernellist}
\title{getkernellist}
\usage{
getkernellist(cValues, tree, OTU, verbose)
}
\arguments{
\item{cValues}{c vector}

\item{tree}{the phylogenetic tree}

\item{OTU}{the OTU table}
}
\description{
getkernellist
}
\keyword{internal}
