% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.post_array}
\alias{plot.post_table}
\title{Plot post_array with plot.stars default}
\usage{
\method{plot}{post_array}(x, y, ...)

\method{plot}{post_table}(x, y, ...)
}
\arguments{
\item{x}{an object of class post_array pr post_table}

\item{y}{ignored}

\item{...}{passed on to \code{plot.stars()}}
}
\description{
Plots a post_array with the default \code{plot.stars()} function by taking the first
non-spatial attribute. The summary geometry is used for plotting.
When no non-spatial attribute is available, the plot function returns an error.
Try \code{autoplot.post_array()} to achieve those type of plots and to plot changing
geometries.
}
\examples{
library(sf)
arr = as_post_array(polygons)
arr$area = st_area(arr$geometry)
plot(arr)

}
\seealso{
\code{\link[=autoplot.post_array]{autoplot.post_array()}}
}
