% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polygons}
\alias{polygons}
\title{Random polygons to exemplify {post} functions}
\format{
An object of class \link{sf} with \code{POLYGON} geometries,
with 25 rows and 3 columns:
\describe{
\item{gid}{ID of polygon geometry}
\item{tid}{ID of polygon timestamp}
\item{geometry}{the geometry list column}
}
}
\usage{
polygons
}
\description{
Five random polygons with 5 timestamp states are artificially
generated to exemplify the main funcitonality of post.
See \link{data-raw/polygons.R} for code on its creation.
}
\keyword{datasets}
