\name{sampExprSet}
\alias{sampExprSet}
 \title{Example of Expression Set}
 \description{an ExpressionSet class contains minimum of exprs (expression matrix) of gene expression and 
              phenoData (AnnotatedDataFrame of end point data).}
 \usage{data(sampExprSet)}
 \value{an simulated ExpressionSet contains conceptual data of 60 expression features for 40 subjects. 
        The phenotype data has 4 columns for the same 40 subjects.}
 \keyword{misc}
