% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf-post_table}
\alias{sf-post_table}
\alias{st_as_sf.post_table}
\title{sf methods for post_table objects}
\usage{
\method{st_as_sf}{post_table}(x, ...)
}
\arguments{
\item{x}{a post_table object}

\item{...}{additional arguments passed onto the respective sf function}
}
\description{
sf methods for post_table objects
}
\details{
\code{sf::st_as_sf()} for post_table objects sets the changing geometry as the
active sf_column but keeps the summarised geometry in a separate column.
}
