% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{stars}
\alias{stars}
\alias{[.post_array}
\title{stars methods for post_array objects}
\usage{
\method{[}{post_array}(x, ...)
}
\arguments{
\item{x}{a post_array object}

\item{...}{additional arguments passed onto the respective stars function}
}
\description{
stars methods for post_array objects
}
