% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_geometry.R
\name{summarise_geometry}
\alias{summarise_geometry}
\alias{summarise_geometry_union}
\alias{summarize_geometry_union}
\alias{summarise_geometry_centroid}
\alias{summarize_geometry_centroid}
\alias{summarise_geometry_bbox}
\alias{summarize_geometry_bbox}
\alias{summarise_geometry_convex_hull}
\alias{summarize_geometry_convex_hull}
\title{Summarise changing geometries based on a group identifier}
\usage{
summarise_geometry_union(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  return_sf = FALSE,
  .checks = TRUE
)

summarize_geometry_union(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  return_sf = FALSE,
  .checks = TRUE
)

summarise_geometry_centroid(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  return_sf = FALSE,
  .checks = TRUE
)

summarize_geometry_centroid(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  return_sf = FALSE,
  .checks = TRUE
)

summarise_geometry_bbox(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  return_sf = FALSE,
  .checks = TRUE,
  rotated = FALSE
)

summarize_geometry_bbox(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  return_sf = FALSE,
  .checks = TRUE,
  rotated = FALSE
)

summarise_geometry_convex_hull(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  return_sf = FALSE,
  .checks = TRUE
)

summarize_geometry_convex_hull(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  return_sf = FALSE,
  .checks = TRUE
)
}
\arguments{
\item{x}{object \code{POLYGON}/\code{MULTIPOLYGON} changing geometries to summarise
based on \code{group_id}.}

\item{group_id}{see \code{?post_array} for details.
Defaults to the first non-spatial, non-temporal column in x.}

\item{sf_column_name}{(character) name of column with changing
geometries. Defaults to active \code{sf_column}.}

\item{return_sf}{should a sf be returned with the group_ids as a column?}

\item{.checks}{internal, should creation arguments be checked?}

\item{rotated}{(logical) if TRUE, the minimum rotated rectangle is returned,
if FALSE the minimum unrotated rectangle or bounding box is returned.
Defaults to FALSE}
}
\value{
object of class \code{sfc} with summarised geometries based on group_id
}
\description{
Changing geometries are summarised for all time periods as a single geometry
(of types \code{POINT}, \code{POLYGON}, \code{MULTIPOLYGON}) in order to pass them
as a spatial dimension to \verb{post_*} objects.
The family of functions \verb{summarise_geometry_*()} or \verb{summarize_geometry_*()}
are helpers to create these summarised geometries.
}
\details{
A custom function to summarise geometries can be passed onto the
coercion functions \verb{as_post_*()}, given that the parameters \code{x},
\code{group_id} and \code{sf_column_name} are included in the function.
Additional arguments can be passed to the function as necessary.
See \code{vignette("post02_geomsum")} for more information.
}
\section{Functions}{
\itemize{
\item \code{summarise_geometry_union()}: computes the geometry summary as the union and
dissolve of the changing geometries

\item \code{summarise_geometry_centroid()}: computes the geometry summary as the centroid
of the union and dissolve of the changing geometries

\item \code{summarise_geometry_bbox()}: computes the geometry summary as the bounding box or
minimum rectangle of the union and dissolve of the changing geometries

\item \code{summarise_geometry_convex_hull()}: computes the geometry summary as the
convex hull of the union and dissolve of the changing geometries

}}
\examples{
# union and dissolve
summarise_geometry_union(polygons)
# centroid
summarise_geometry_centroid(polygons)
# bounding box or minimum unrotated rectangle
summarise_geometry_bbox(polygons)
# minimum rotated rectangle
summarise_geometry_bbox(polygons, rotated = TRUE)
# convex hull
summarise_geometry_convex_hull(polygons)

}
\concept{Changing geometries summarisers}
