% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_array.R
\name{utils-post-array}
\alias{utils-post-array}
\alias{get_group_ids}
\title{Utility functions for post_array objects}
\usage{
get_group_ids(x)
}
\arguments{
\item{x}{a post_array object}
}
\value{
a data.frame with the group_id values
}
\description{
Utility functions for post_array objects
}
\details{
\code{get_group_ids()} gets the group_id values of the post_array object as a
named data.frame
}
