% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_post.R
\name{as_post_array}
\alias{as_post_array}
\alias{as_post_array.sf}
\title{Create a post_array object}
\usage{
as_post_array(
  x,
  group_id,
  time_column_name,
  sf_column_name = attr(x, "sf_column"),
  geometry_summary = "centroid",
  attributes = NULL
)

\method{as_post_array}{sf}(
  x,
  group_id,
  time_column_name,
  sf_column_name = attr(x, "sf_column"),
  geometry_summary = "centroid",
  attributes = NULL
)
}
\arguments{
\item{x}{object of class \code{sf} with \code{POLYGON} geometries}

\item{group_id}{(character) name of column containing the grouping
identifier for the changing polygons.}

\item{time_column_name}{(character) name of column with the datetimes}

\item{sf_column_name}{(character) name of column with changing
geometries. Defaults to active \code{sf_column}.}

\item{geometry_summary}{(character) function to compute the summary
geometry. Supported arguments are \code{union} to compute the union
and dissolve of the changing geometries per group, \code{centroid} to
compute the centroid of the union and dissolve per group, and
\code{bbox} to compute the bbox of the union and dissolve per group.}

\item{attributes}{any other attribute in the \code{sf} object that
should be passed onto the \code{post_array} object.}
}
\value{
an object of class \code{post_array}.
}
\description{
\code{post_array} is a spatio-temporal array data structure that
organises polygons that change their shape in space and time.
It extends the \code{stars} class for vector data cubes to support
changing geometries as attributes.
The \code{post_array} class supports two dimensions: \code{geom_sum} and
\code{datetime}. \code{geom_sum} is the summary geometry of the changing
shapes of the polygon geometries.
}
\examples{
as_post_array(polygons, "gid", "datetime")

library(sf, quietly = TRUE)
polygons |>
  mutate(area = st_area(geometry)) |>
  as_post_array("gid", "datetime", attributes = "area")
}
\references{
Abad, L., Sudmanns, M., and Hölbling, D. (2024)
Vector data cubes for features evolving in space and time,
\emph{AGILE GIScience Ser.}, 5, 16,
\url{https://doi.org/10.5194/agile-giss-5-16-2024}
}
