% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_array.R
\name{as_post_array}
\alias{as_post_array}
\alias{as_post_array.sf}
\alias{as_post_array.post_table}
\alias{as_post_array.post_array}
\title{Create a post_array object}
\usage{
as_post_array(
  x,
  group_id = NULL,
  time_column_name = NULL,
  sf_column_name = NULL,
  geometry_summary = summarise_geometry_centroid,
  geometry_summary_name = NULL,
  ...
)

\method{as_post_array}{sf}(
  x,
  group_id = NULL,
  time_column_name = NULL,
  sf_column_name = NULL,
  geometry_summary = summarise_geometry_centroid,
  geometry_summary_name = NULL,
  ...,
  point_st = c(TRUE, FALSE)
)

\method{as_post_array}{post_table}(x, ...)

\method{as_post_array}{post_array}(x, ...)
}
\arguments{
\item{x}{object to convert to \code{post_array} with \code{POLYGON}/\code{MULTIPOLYGON}
geometries and a date/time column.}

\item{group_id}{(character) name of column containing the grouping
identifier for the changing polygons or group identifier given as an
integer or character value that is assigned to all rows in \code{x}.
Defaults to the first column in \code{x}.}

\item{time_column_name}{(character) name of column with the temporal
dimension information. Defaults to the first temporal column in \code{x}.}

\item{sf_column_name}{(character) name of column with changing
geometries. Defaults to active \code{sf_column}.}

\item{geometry_summary}{(function) function to compute the summary
geometry. Alternatively an \code{sfc} object with summary geometries (i.e.,
the result of a \verb{summarise_geometry_*()} function or a custom \code{sfc}).
See summarise_geometry for functions or pass your own
summarise_geometry function.}

\item{geometry_summary_name}{(character) name of the column with summary
geometries.Defaults to "geom_sum"}

\item{...}{additional parameters passed on to \code{geometry_summary} function.}

\item{point_st}{point argument passed onto \code{stars::st_dimensions()}.
Defaults to \code{TRUE} for the spatial dimension and to \code{FALSE} for the
temporal dimension (assumes interval times)}
}
\value{
an object of class \code{post_array}.
}
\description{
\code{post_array} is a spatio-temporal array data structure that
organises polygons that change their shape in space and time.
It extends the \code{stars} class for vector data cubes to support
changing geometries as attributes.
The \code{post_array} class supports two dimensions: \code{geom_sum} and
\code{datetime}.
\code{geom_sum} is the summary geometry of the changing
shapes of the polygon geometries.
}
\examples{
as_post_array(polygons)

if(require(dplyr, quietly = TRUE)) {
  library(sf, quietly = TRUE)
  polygons |>
    mutate(area = st_area(geometry)) |>
    as_post_array(
     geometry_summary = summarise_geometry_union
   )
}

}
\references{
Abad, L., Sudmanns, M., and Hölbling, D. (2024)
Vector data cubes for features evolving in space and time,
\emph{AGILE GIScience Ser.}, 5, 16,
\url{https://doi.org/10.5194/agile-giss-5-16-2024}
}
