% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble.R
\name{cubble-fill}
\alias{cubble-fill}
\alias{fill_gaps.post_table}
\title{Gap-filling on cubble}
\usage{
\method{fill_gaps}{post_table}(.data)
}
\arguments{
\item{data}{a post_table object}
}
\value{
a post_table object
}
\description{
Gap-filling on cubble
}
