% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf-post-array}
\alias{sf-post-array}
\alias{st_as_sf.post_array}
\title{sf methods for post_array objects}
\usage{
\method{st_as_sf}{post_array}(x)
}
\arguments{
\item{x}{a post_array object}
}
\description{
sf methods for post_array objects
}
\details{
\code{sf::st_as_sf()} for post_array objects sets the changing geometry as the
active sf_column but keeps the summarised geometry in a separate column.
}
