% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_geometry.R
\name{summarise_geometry}
\alias{summarise_geometry}
\alias{summarise_geometry_union}
\alias{summarize_geometry_union}
\alias{summarise_geometry_centroid}
\alias{summarize_geometry_centroid}
\alias{summarise_geometry_bbox}
\alias{summarize_geometry_bbox}
\title{Summarise changing geometries based on a group identifier}
\usage{
summarise_geometry_union(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  .checks = TRUE
)

summarize_geometry_union(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  .checks = TRUE
)

summarise_geometry_centroid(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  .checks = TRUE
)

summarize_geometry_centroid(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  .checks = TRUE
)

summarise_geometry_bbox(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  .checks = TRUE
)

summarize_geometry_bbox(
  x,
  group_id = NULL,
  sf_column_name = NULL,
  .checks = TRUE
)
}
\arguments{
\item{x}{object \code{POLYGON}/\code{MULTIPOLYGON} changing geometries to summarise
based on \code{group_id}.}

\item{group_id}{See Details.
Defaults to the first column in \code{x}.}

\item{sf_column_name}{(character) name of column with changing
geometries. Defaults to active \code{sf_column}.}

\item{.checks}{perform argument check?}
}
\value{
object of class \code{sfc} with summarised geometries based on group_id
}
\description{
Changing geometries are summarised for all time periods as a single geometry
(of types \code{POINT}, \code{POLYGON}, \code{MULTIPOLYGON}) in order to pass them
as a spatial dimension to \verb{post_*} objects.

\code{summarise_geometry_union} computes the geometry summary as the union and dissolve of
the changing geometries

\code{summarise_geometry_centroid} computes the geometry summary as the centroid of the
union and dissolve of the changing geometries

\code{summarise_geometry_bbox} computes the geometry summary as the bounding box of the
union and dissolve of the changing geometries
}
\details{
The family of functions \verb{summarise_geometry_*()} or \verb{summarize_geometry_*()}
are helpers to create these summarised geometries.
Supported summary types are:

a. the union and dissolve of the changing geometries,

b. the centroid of \strong{a}

c. the bbox of \strong{a}

A custom function to summarise geometries can be passed onto the
coercion functions \verb{as_post_*()}, given that the parameters \code{x},
\code{group_id} and \code{sf_column_name} are included in the function.
Additional arguments can be passed to the function as necessary.
}
\examples{
summarise_geometry_union(polygons)
summarise_geometry_centroid(polygons)
summarise_geometry_bbox(polygons)

}
\concept{Changing geometries summarisers}
